feat: enhance UserService with resource not found checks and improve exception message format

+++ src/main/java/com/spproject/springservices/services/UserService.java
@@ -11,40 +11,43 @@ import com.spproject.springservices.services.exceptions.ResourceNotFoundExceptio
 
 @Service
 public class UserService {
-    
+
     private final UserRepository repository;
 
     public UserService(UserRepository repository) {
         this.repository = repository;
     }
-    
-    public List<User> findAll(){
+
+    public List<User> findAll() {
         return repository.findAll();
     }
 
-    public User findById(Long id){
+    public User findById(Long id) {
         Optional<User> obj = repository.findById(id);
         return obj.orElseThrow(() -> new ResourceNotFoundException(id));
     }
 
-    public User insert(User obj){
+    public User insert(User obj) {
         return repository.save(obj);
     }
 
-    public void delete(Long id){
+    public void delete(Long id) {
+        if (!repository.existsById(id)) {
+            throw new ResourceNotFoundException(id);
+        }
         repository.deleteById(id);
     }
 
-    public User update (Long id, User obj){
+    public User update(Long id, User obj) {
         User entity = repository.getReferenceById(id);
         updateData(entity, obj);
-                return repository.save(entity);
-            }
-        
-            private void updateData(User entity, User obj) {
-                entity.setName(obj.getName());
-                entity.setEmail(obj.getEmail());
-                entity.setPhone(obj.getPhone());
-            }
+        return repository.save(entity);
+    }
+
+    private void updateData(User entity, User obj) {
+        entity.setName(obj.getName());
+        entity.setEmail(obj.getEmail());
+        entity.setPhone(obj.getPhone());
+    }
 
 }
